fix(oba): correct the JOC sparse differential and GCC sign-conversion warnings
6.6.2 Pseudocode 2 codes each band's coefficient against
joc_mix_mtx_q[ch][pb-1] for the channel THAT band names - which is the
previous band's decoded value only when that band named the same channel,
and the sparse offset otherwise, since that is what every unnamed channel
holds. A single running previous carried a value across a channel change
that the clause never puts there.

The rest is -Wsign-conversion under GCC, which MSVC does not diagnose.

+++ src/forge/include/ac3/oba/oamd.hpp
@@ -206,8 +206,8 @@ struct Program {
 // count is just a popcount - unlike Table 12, where a set bit can name a pair.
 [[nodiscard]] constexpr int nonstd_bed_channel_count(std::uint32_t assignment) {
     int count = 0;
-    for (int bit = 0; bit < 17; ++bit) {
-        count += (assignment >> bit) & 1u;
+    for (unsigned bit = 0; bit < 17U; ++bit) {
+        count += static_cast<int>((assignment >> bit) & 1U);
     }
     return count;
 }
